no jvm or android target (for testing purposes)
Let's check out if without android nor JVM as target, if we can still use this module as dependency in our Android application. My guess: the metadata.jar isn't sufficient, we still need jvm (not Android, as we don't have any resources or UI => so no .aar needed) as target.